# docs/swagger.yaml
basePath: /api
definitions:
  models.HealthCheckConfigAPI:
    properties:
      enabled:
        type: boolean
      endpoint:
        type: string
      healthy_threshold:
        type: integer
      interval:
        description: Duration as string
        type: string
      timeout:
        description: Duration as string
        type: string
      unhealthy_threshold:
        type: integer
    type: object
  models.LoadBalancingConfigAPI:
    properties:
      hash_key:
        type: string
      session_key:
        type: string
      sticky_session:
        type: boolean
      strategy:
        type: string
    type: object
  models.RouteDestinationAPI:
    properties:
      broker_queue:
        type: string
      config:
        additionalProperties: true
        type: object
      health_check:
        $ref: '#/definitions/models.HealthCheckConfigAPI'
      id:
        type: string
      pipeline_id:
        type: string
      priority:
        type: integer
      timeout:
        description: Duration as string
        type: string
      type:
        type: string
      url:
        type: string
      weight:
        type: integer
    type: object
  models.RouteRuleAPI:
    properties:
      conditions:
        items:
          $ref: '#/definitions/models.RuleConditionAPI'
        type: array
      created_at:
        type: string
      description:
        type: string
      destinations:
        items:
          $ref: '#/definitions/models.RouteDestinationAPI'
        type: array
      enabled:
        type: boolean
      id:
        type: string
      load_balancing:
        $ref: '#/definitions/models.LoadBalancingConfigAPI'
      name:
        type: string
      priority:
        type: integer
      tags:
        items:
          type: string
        type: array
      updated_at:
        type: string
    type: object
  models.RuleConditionAPI:
    properties:
      field:
        type: string
      negate:
        type: boolean
      operator:
        type: string
      type:
        type: string
      value: {}
    type: object
...
  routing.RouteRule:
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Enterprise-grade webhook routing and data transformation platform
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Webhook Router API
  version: "1.0"
paths:
...
  /routing/rules:
    get:
      description: Returns a list of all configured advanced routing rules
      produces:
      - application/json
      responses:
        "200":
          description: List of routing rules
          schema:
            items:
              $ref: '#/definitions/models.RouteRuleAPI'
            type: array
        "500":
          description: Internal server error
          schema:
            type: string
        "503":
          description: Router not initialized
          schema:
            type: string
      security:
      - SessionAuth: []
      summary: Get all routing rules
      tags:
      - routing
    post:
      consumes:
      - application/json
      description: Creates a new advanced routing rule with conditions and destinations
      parameters:
      - description: Routing rule configuration
        in: body
        name: rule
        required: true
        schema:
          $ref: '#/definitions/routing.RouteRule'
      produces:
      - application/json
      responses:
        "201":
          description: Created routing rule
          schema:
            $ref: '#/definitions/routing.RouteRule'
        "400":
          description: Invalid JSON or rule validation failed
          schema:
            type: string
        "503":
          description: Router not initialized
          schema:
            type: string
      security:
      - SessionAuth: []
      summary: Create routing rule
      tags:
      - routing
  /routing/rules/{id}:
    delete:
      description: Removes a routing rule configuration
      parameters:
      - description: Rule ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Rule not found
          schema:
            type: string
        "503":
          description: Router not initialized
          schema:
            type: string
      security:
      - SessionAuth: []
      summary: Delete routing rule
      tags:
      - routing
    get:
      description: Returns a specific routing rule by ID
      parameters:
      - description: Rule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Routing rule
          schema:
            $ref: '#/definitions/models.RouteRuleAPI'
        "404":
          description: Rule not found
          schema:
            type: string
        "503":
          description: Router not initialized
          schema:
            type: string
      security:
      - SessionAuth: []
      summary: Get routing rule
      tags:
      - routing
    put:
      consumes:
      - application/json
      description: Updates an existing routing rule configuration
      parameters:
      - description: Rule ID
        in: path
        name: id
        required: true
        type: string
      - description: Routing rule configuration
        in: body
        name: rule
        required: true
        schema:
          $ref: '#/definitions/routing.RouteRule'
      produces:
      - application/json
      responses:
        "200":
          description: Updated routing rule
          schema:
            $ref: '#/definitions/routing.RouteRule'
        "400":
          description: Invalid JSON or rule validation failed
          schema:
            type: string
        "503":
          description: Router not initialized
          schema:
            type: string
      security:
      - SessionAuth: []
      summary: Update routing rule
      tags:
      - routing
  /routing/rules/{id}/test:
    post:
      consumes:
      - application/json
      description: Tests a routing rule with sample request data
      parameters:
      - description: Rule ID
        in: path
        name: id
        required: true
        type: string
      - description: Test request data
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Test result
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid JSON
          schema:
            type: string
        "404":
          description: Rule not found
          schema:
            type: string
        "500":
          description: Route test failed
          schema:
            type: string
        "503":
          description: Router not initialized
          schema:
            type: string
      security:
      - SessionAuth: []
      summary: Test routing rule
      tags:
      - routing
...
securityDefinitions:
  SessionAuth:
    in: cookie
    name: session
    type: apiKey
swagger: "2.0"